+++ src/main/java/com/plainprog/grandslam_ai/entity/account_security/AccountSecurityRepository.java
@@ -2,7 +2,7 @@ package com.plainprog.grandslam_ai.entity.account_security;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountSecurityRepository extends JpaRepository<AccountSecurity, Integer> {
-    public AccountSecurity findByAccountId(Integer accountId);
+public interface AccountSecurityRepository extends JpaRepository<AccountSecurity, Long> {
+    public AccountSecurity findByAccountId(Long accountId);
     public AccountSecurity findByVerifyEmailToken(String verifyEmailToken);
 }
