+++ src/main/java/com/plainprog/grandslam_ai/entity/img_gen/ImageRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ImageRepository extends JpaRepository<Image, Integer> {
-    Optional<Image> findFirstByOwnerAccountId(Integer ownerAccountId);
-    Optional<Image> findFirstByOwnerAccountIdNot(Integer id);
+public interface ImageRepository extends JpaRepository<Image, Long> {
+    Optional<Image> findFirstByOwnerAccountId(Long ownerAccountId);
+    Optional<Image> findFirstByOwnerAccountIdNot(Long id);
 }
